+++ docs/changelog.md
@@ -39,4 +39,19 @@
 - **Authentication Services**
   Added necessary authentication services in `src/services/auth_services.ts`
 
+### 2025-11-23(SF)
 
+### Progress
+- **Authentication Services**
+  Added necessary authentication services in
+  `src/services/spotify_auth.ts`
+  - export async function linkSpotifyAccount() { ... }
+  - export async function unlinkSpotifyAccount() { ... }
+
+### 2025-11-23(Eric)
+
+### Progress
+- **Spotify Services**
+  Added extra spotify services
+  - export async function createSpotifyPlaylist(userId, name, desc, isPublic) { ... }
+  - export async function addTracksToSpotifyPlaylist(playlistId, trackUris) { ... }